% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Código TCC.R
\name{Esc_perf}
\alias{Esc_perf}
\title{Função para dimensionar perfis de aço laminado (flexão e cortante)}
\usage{
Esc_perf(Msd_Solicitante, Vsd_Solicitante, Lb, Cb)
}
\arguments{
\item{Msd_Solicitante}{Momento fletor solicitante de cálculo (em N.mm)}

\item{Vsd_Solicitante}{Força cortante solicitante de cálculo (em N)}

\item{Lb}{Comprimento destravado da estrutura (em mm)}

\item{Cb}{Fator de modificação para diagrama de momento fletor não-uniforme (caso não haja, adotar igual a 1.0)}
}
\description{
Essa função serve para dimensionar perfis de aço laminado sujeitos a momentos fletores e esforços cortantes
}
\examples{
Esc_perf(231240000,116400,3200,1)

}
